+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- **BREAKING:** Namespace changed from `nikolapopovic\angiechat` to `Dz0nika\AngieChatCraft` to align with the Composer package name `dz0nika/angie-chat-craft`. If you have extended any plugin classes, update your `use` statements accordingly.
+
 ## [1.0.0] - 2026-03-04
 
 ### Added